requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
flake8==3.5.0
gunicorn==19.9.0
mccabe==0.6.1
packaging==17.1
pbr==4.2.0
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytz==2018.5
six==1.11.0
stevedore==1.29.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
whitenoise==4.1
django-bootstrap4==0.0.7
django-rest-swagger==2.2.0
djangorestframework==3.9.0
django-icons==0.2.1
django-rest-auth==0.9.3
django-polymorphic==2.0.3
django-rest-polymorphic==0.1.8
